+++ types/assert/package.json
@@ -2,9 +2,9 @@
     "private": true,
     "types": "index",
     "typesVersions": {
-        ">=3.7.0-0": {
+        "<=3.6": {
             "*": [
-                "ts3.7/*"
+                "ts3.6/*"
             ]
         }
     }
