+++ types/assert/ts3.6/assert-tests.ts
@@ -1,0 +1,19 @@
+import * as assert from 'assert';
+
+assert(true, "it's working");
+
+assert.ok(true, 'inner functions work as well');
+
+assert.throws(() => {});
+assert.throws(() => {}, /Regex test/);
+assert.throws(
+    () => {},
+    () => {},
+    'works wonderfully',
+);
+
+assert['fail'](true, true, 'works like a charm');
+
+assert.strict; // $ExpectType typeof assert
+
+assert.fail(); // $ExpectType never
