+++ types/assert/ts3.6/index.d.ts
@@ -1,27 +1,23 @@
-declare function assert(value: any, message?: string): asserts value;
+declare function assert(value: any, message?: string): void;
 
 declare namespace assert {
     function fail(actual?: any, expected?: any, message?: string, operator?: string): never;
 
-    function ok(value: any, message?: string): asserts value;
+    function ok(value: any, message?: string): void;
 
-    /** @deprecated Use `strictEqual` instead */
     function equal(actual: any, expected: any, message?: string): void;
 
-    /** @deprecated Use `notStrictEqual` instead */
     function notEqual(actual: any, expected: any, message?: string): void;
 
-    /** @deprecated Use `deepStrictEqual` instead */
     function deepEqual(actual: any, expected: any, message?: string): void;
 
-    /** @deprecated Use `notDeepStrictEqual` instead */
     function notDeepEqual(actual: any, expected: any, message?: string): void;
 
-    function deepStrictEqual<T>(actual: any, expected: T, message?: string): asserts actual is T;
+    function deepStrictEqual(actual: any, expected: any, message?: string): void;
 
     function notDeepStrictEqual(actual: any, expected: any, message?: string): void;
 
-    function strictEqual<T>(actual: any, expected: T, message?: string): asserts actual is T;
+    function strictEqual(actual: any, expected: any, message?: string): void;
 
     function notStrictEqual(actual: any, expected: any, message?: string): void;
 
@@ -35,7 +31,7 @@ declare namespace assert {
         message?: string,
     ): void;
 
-    function ifError(value: any): asserts value is null | undefined;
+    function ifError(value: any): void;
 
     class AssertionError implements Error {
         name: string;
